Add (horribly-implemented) buzzer
The buzzer is a sine wave that's being pumped to SDL. It's pretty
crusty, though. I believe that I should not be generating the sine wave
from inside the main render/emulation loop the way that I am. Maybe a
callback function to fill the waveform buffer would produce better
output...